DICompileUnits are always distinct
See link included in the comment for details

TODO: Does this break older versions in any situation? There doesn't
seem to be an easy way to tell from the information in the record if it
was output by llvm-as of a version later than the linked commit.

One way to definitely NOT break really old versions, but perhaps break
versions between that commit and LLVM 5.0 would be to set:

  isDistinct = Record[0] || (getModVersion >= 2)

Then it would be always-true for bitcode produced after LLVM 5.

This sort of issue highlights the need for a larger test corpus...